+++ core/shared/src/test/scala/fs2/PullLawsSuite.scala
@@ -8,26 +8,8 @@ import cats.effect.laws.util.TestInstances._
 import cats.implicits._
 import cats.effect.laws.discipline._
 
-import org.scalacheck.{Arbitrary, Gen}
-import Arbitrary.arbitrary
-
-class PullLawsSpec extends Fs2Spec with StreamArbitrary {
+class PullLawsSuite extends Fs2Suite {
   implicit val ec: TestContext = TestContext()
-
-  implicit def arbPull[F[_], O, R](implicit
-      arbR: Arbitrary[R],
-      arbFr: Arbitrary[F[R]],
-      arbFo: Arbitrary[F[O]],
-      arbO: Arbitrary[O],
-      arbFu: Arbitrary[F[Unit]]
-  ): Arbitrary[Pull[F, O, R]] =
-    Arbitrary(
-      Gen.oneOf(
-        arbitrary[R].map(Pull.pure(_)),
-        arbitrary[F[R]].map(Pull.eval(_)),
-        arbitrary[Stream[F, O]].flatMap(s => arbitrary[R].map(r => s.pull.echo >> Pull.pure(r)))
-      )
-    )
 
   implicit def eqPull[O: Eq, R: Eq]: Eq[Pull[IO, O, R]] = {
     def normalize(p: Pull[IO, O, R]): IO[Vector[Either[O, R]]] =
