+++ core/shared/src/test/scala/fs2/StreamLawsSuite.scala
@@ -9,30 +9,7 @@ import cats.implicits._
 import cats.laws.discipline._
 import cats.laws.discipline.arbitrary._
 
-import org.scalacheck.{Arbitrary, Gen}
-import Arbitrary.arbitrary
-
-trait StreamArbitrary {
-  implicit def arbStream[F[_], O](implicit
-      arbO: Arbitrary[O],
-      arbFo: Arbitrary[F[O]],
-      arbFu: Arbitrary[F[Unit]]
-  ): Arbitrary[Stream[F, O]] =
-    Arbitrary(
-      Gen.frequency(
-        10 -> arbitrary[List[O]].map(os => Stream.emits(os).take(10)),
-        10 -> arbitrary[List[O]].map(os => Stream.emits(os).take(10).unchunk),
-        5 -> arbitrary[F[O]].map(fo => Stream.eval(fo)),
-        1 -> (for {
-          acquire <- arbitrary[F[O]]
-          release <- arbitrary[F[Unit]]
-          use <- arbStream[F, O].arbitrary
-        } yield Stream.bracket(acquire)(_ => release).flatMap(_ => use))
-      )
-    )
-}
-
-class StreamLawsSpec extends Fs2Spec with StreamArbitrary {
+class StreamLawsSuite extends Fs2Suite {
   implicit val ec: TestContext = TestContext()
 
   implicit def eqStream[O: Eq]: Eq[Stream[IO, O]] =
