+++ core/shared/src/test/scala/fs2/StreamSpec.scala
@@ -17,765 +17,6 @@ class StreamSpec extends Fs2Spec {
 
   "Stream" - {
 
-    "cancelation of compiled streams" - {
-      def startAndCancelSoonAfter[A](fa: IO[A]): IO[Unit] =
-        fa.start.flatMap(fiber => IO.sleep(1.second) >> fiber.cancel)
-
-      def testCancelation[A](s: Stream[IO, A]): IO[Assertion] =
-        startAndCancelSoonAfter(s.compile.drain).assertNoException
-
-      def constantStream: Stream[IO, Int] =
-        if (isJVM) Stream.constant(1) else Stream.constant(1).evalTap(_ => IO.sleep(1.milli))
-
-      "constant" in testCancelation(constantStream)
-
-      "bracketed stream" in testCancelation(
-        Stream.bracket(IO.unit)(_ => IO.unit).flatMap(_ => constantStream)
-      )
-
-      "concurrently" in testCancelation {
-        constantStream.concurrently(constantStream)
-      }
-
-      "merge" in testCancelation {
-        constantStream.merge(constantStream)
-      }
-
-      "parJoin" in testCancelation {
-        Stream(constantStream, constantStream).parJoin(2)
-      }
-
-      "#1236" in testCancelation {
-        Stream
-          .eval(Queue.bounded[IO, Int](1))
-          .flatMap { q =>
-            Stream(
-              Stream
-                .unfold(0)(i => (i + 1, i + 1).some)
-                .flatMap(i => Stream.sleep_(50.milliseconds) ++ Stream.emit(i))
-                .through(q.enqueue),
-              q.dequeue.drain
-            ).parJoin(2)
-          }
-      }
-    }
-
-    "changes" in {
-      assert(Stream.empty.covaryOutput[Int].changes.toList == Nil)
-      assert(Stream(1, 2, 3, 4).changes.toList == List(1, 2, 3, 4))
-      assert(Stream(1, 1, 2, 2, 3, 3, 4, 3).changes.toList == List(1, 2, 3, 4, 3))
-      val result = Stream("1", "2", "33", "44", "5", "66")
-        .changesBy(_.length)
-        .toList
-      assert(result == List("1", "33", "5", "66"))
-    }
-
-    "chunkLimit" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-      val n = n0 % 20 + 1
-      val sizeV = s.chunkLimit(n).toVector.map(_.size)
-      assert(sizeV.forall(_ <= n))
-      assert(sizeV.combineAll == s.toVector.size)
-    }
-
-    "chunkMin" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-      val n = n0 % 20 + 1
-      val chunkedV = s.chunkMin(n, true).toVector
-      val withIfSmallerV = s.chunkMin(n, false).toVector
-      val unchunkedV = s.toVector
-      val smallerSet = s.take(n - 1).toVector
-      val smallerN = s.take(n - 1).chunkMin(n, false).toVector
-      val smallerY = s.take(n - 1).chunkMin(n, true).toVector
-      // All but last list have n values
-      assert(chunkedV.dropRight(1).forall(_.size >= n))
-      // Equivalent to last chunk with allowFewerTotal
-      if (chunkedV.nonEmpty && chunkedV.last.size < n)
-        assert(chunkedV.dropRight(1) == withIfSmallerV)
-      // Flattened sequence with allowFewerTotal true is equal to vector without chunking
-      assert(chunkedV.foldLeft(Vector.empty[Int])((v, l) => v ++ l.toVector) == unchunkedV)
-      // If smaller than Chunk Size and allowFewerTotal false is empty then
-      // no elements should be emitted
-      assert(smallerN == Vector.empty)
-      // If smaller than Chunk Size and allowFewerTotal true is equal to the size
-      // of the taken chunk initially
-      assert(smallerY.foldLeft(Vector.empty[Int])((v, l) => v ++ l.toVector) == smallerSet)
-    }
-
-    "chunkN" - {
-      "fewer" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-        val n = n0 % 20 + 1
-        val chunkedV = s.chunkN(n, true).toVector
-        val unchunkedV = s.toVector
-        // All but last list have n0 values
-        assert(chunkedV.dropRight(1).forall(_.size == n))
-        // Last list has at most n0 values
-        assert(chunkedV.lastOption.fold(true)(_.size <= n))
-        // Flattened sequence is equal to vector without chunking
-        assert(chunkedV.foldLeft(Vector.empty[Int])((v, l) => v ++ l.toVector) == unchunkedV)
-      }
-
-      "no-fewer" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
-        val n = n0 % 20 + 1
-        val chunkedV = s.chunkN(n, false).toVector
-        val unchunkedV = s.toVector
-        val expectedSize = unchunkedV.size - (unchunkedV.size % n)
-        // All lists have n0 values
-        assert(chunkedV.forall(_.size == n))
-        // Flattened sequence is equal to vector without chunking, minus "left over" values that could not fit in a chunk
-        val left = chunkedV.foldLeft(Vector.empty[Int])((v, l) => v ++ l.toVector)
-        val right = unchunkedV.take(expectedSize)
-        assert(left == right)
-      }
-    }
-
-    "chunks" - {
-      "chunks.map identity" in forAll { (v: Vector[Vector[Int]]) =>
-        val s = if (v.isEmpty) Stream.empty else v.map(Stream.emits).reduce(_ ++ _)
-        assert(s.chunks.map(_.toVector).toVector == v.filter(_.nonEmpty))
-      }
-
-      "chunks.flatMap(chunk) identity" in forAll { (v: Vector[Vector[Int]]) =>
-        val s = if (v.isEmpty) Stream.empty else v.map(Stream.emits).reduce(_ ++ _)
-        assert(s.chunks.flatMap(Stream.chunk).toVector == v.flatten)
-      }
-    }
-
-    "collect" in forAll { (s: Stream[Pure, Int]) =>
-      val pf: PartialFunction[Int, Int] = { case x if x % 2 == 0 => x }
-      assert(s.collect(pf).toVector == s.toVector.collect(pf))
-    }
-
-    "collectFirst" in forAll { (s: Stream[Pure, Int]) =>
-      val pf: PartialFunction[Int, Int] = { case x if x % 2 == 0 => x }
-      assert(s.collectFirst(pf).toVector.headOption == s.toVector.collectFirst(pf))
-    }
-
-    "collectWhile" in forAll { (s1: Stream[Pure, Int], s2: Stream[Pure, Int]) =>
-      val pf: PartialFunction[Int, Int] = { case x if x % 2 == 0 => x }
-      val even = s1.filter(_ % 2 == 0)
-      val odd = s2.filter(_ % 2 != 0)
-      assert((even ++ odd).collectWhile(pf).toVector == even.toVector)
-    }
-
-    "compile" - {
-      "resource" - {
-        "concurrently" in {
-          val prog: Resource[IO, IO[Unit]] =
-            Stream
-              .eval(Deferred[IO, Unit].product(Deferred[IO, Unit]))
-              .flatMap {
-                case (startCondition, waitForStream) =>
-                  val worker = Stream.eval(startCondition.get) ++ Stream.eval(
-                    waitForStream.complete(())
-                  )
-                  val result = startCondition.complete(()) >> waitForStream.get
-
-                  Stream.emit(result).concurrently(worker)
-              }
-              .compile
-              .resource
-              .lastOrError
-          prog.use(x => x).assertNoException
-        }
-
-        "onFinalise" in {
-          val expected = List(
-            "stream - start",
-            "stream - done",
-            "io - done",
-            "io - start",
-            "resource - start",
-            "resource - done"
-          )
-
-          Ref[IO]
-            .of(List.empty[String])
-            .flatMap { st =>
-              def record(s: String): IO[Unit] = st.update(_ :+ s)
-
-              def stream =
-                Stream
-                  .emit("stream - start")
-                  .onFinalize(record("stream - done"))
-                  .evalMap(x => record(x))
-                  .compile
-                  .lastOrError
-
-              def io =
-                Stream
-                  .emit("io - start")
-                  .onFinalize(record("io - done"))
-                  .compile
-                  .lastOrError
-                  .flatMap(x => record(x))
-
-              def resource =
-                Stream
-                  .emit("resource - start")
-                  .onFinalize(record("resource - done"))
-                  .compile
-                  .resource
-                  .lastOrError
-                  .use(x => record(x))
-
-              stream >> io >> resource >> st.get
-            }
-            .asserting(it => assert(it == expected))
-        }
-
-        "onFinalizeWeak" in {
-          Ref[IO]
-            .of(List.empty[String])
-            .flatMap { st =>
-              def record(s: String): IO[Unit] = st.update(_ :+ s)
-              Stream
-                .emit("emit")
-                .onFinalize(record("1")) // This gets closed
-                .onFinalize(record("2")) // This gets extended
-                .onFinalizeWeak(record("3")) // This joins extended
-                .onFinalizeWeak(record("4")) // This joins extended
-                .compile
-                .resource
-                .lastOrError
-                .use(x => record(x)) >> st.get
-            }
-            .asserting(it => assert(it == List("1", "emit", "2", "3", "4")))
-        }
-
-        "last scope extended, not all scopes" - {
-          "1" in {
-            Ref[IO]
-              .of(List.empty[String])
-              .flatMap { st =>
-                def record(s: String): IO[Unit] = st.update(_ :+ s)
-                Stream
-                  .emit("start")
-                  .onFinalize(record("first finalize"))
-                  .onFinalize(record("second finalize"))
-                  .compile
-                  .resource
-                  .lastOrError
-                  .use(x => record(x)) *> st.get
-              }
-              .asserting(it => assert(it == List("first finalize", "start", "second finalize")))
-          }
-          "2" in {
-            Ref[IO]
-              .of(List.empty[String])
-              .flatMap { st =>
-                def record(s: String): IO[Unit] = st.update(_ :+ s)
-                (Stream.bracket(IO("a"))(_ => record("first finalize")) ++
-                  Stream.bracket(IO("b"))(_ => record("second finalize")) ++
-                  Stream.bracket(IO("c"))(_ =>
-                    record("third finalize")
-                  )).compile.resource.lastOrError
-                  .use(x => record(x)) *> st.get
-              }
-              .asserting(it =>
-                assert(it == List("first finalize", "second finalize", "c", "third finalize"))
-              )
-          }
-        }
-
-        "allocated" in {
-          Ref[IO]
-            .of(false)
-            .flatMap { written =>
-              Stream
-                .emit(())
-                .onFinalize(written.set(true))
-                .compile
-                .resource
-                .lastOrError
-                .allocated >> written.get
-            }
-            .asserting(it => assert(it == false))
-        }
-
-        "interruption" - {
-          "1" in {
-            Stream
-              .resource {
-                Stream.never[IO].compile.resource.drain
-              }
-              .interruptAfter(200.millis)
-              .drain
-              .append(Stream.emit(true))
-              .compile
-              .lastOrError
-              .timeout(2.seconds)
-              .asserting(it => assert(it == true))
-          }
-
-          "2" in {
-            val p = (Deferred[IO, ExitCase[Throwable]]).flatMap { stop =>
-              val r = Stream
-                .never[IO]
-                .compile
-                .resource
-                .drain
-                .use(_ => IO.unit)
-                .guaranteeCase(stop.complete)
-
-              r.start.flatMap(fiber => IO.sleep(200.millis) >> fiber.cancel >> stop.get)
-            }
-            p.timeout(2.seconds)
-              .asserting(it => assert(it == ExitCase.Canceled))
-          }
-        }
-      }
-    }
-
-    "concurrently" - {
-      "when background stream terminates, overall stream continues" in forAll {
-        (s1: Stream[Pure, Int], s2: Stream[Pure, Int]) =>
-          val expected = s1.toList
-          s1.delayBy[IO](25.millis)
-            .concurrently(s2)
-            .compile
-            .toList
-            .asserting(it => assert(it == expected))
-      }
-
-      "when background stream fails, overall stream fails" in forAll { (s: Stream[Pure, Int]) =>
-        s.delayBy[IO](25.millis)
-          .concurrently(Stream.raiseError[IO](new Err))
-          .compile
-          .drain
-          .assertThrows[Err]
-      }
-
-      "when primary stream fails, overall stream fails and background stream is terminated" in {
-        Stream
-          .eval(Semaphore[IO](0))
-          .flatMap { semaphore =>
-            val bg = Stream.repeatEval(IO(1) *> IO.sleep(50.millis)).onFinalize(semaphore.release)
-            val fg = Stream.raiseError[IO](new Err).delayBy(25.millis)
-            fg.concurrently(bg)
-              .onFinalize(semaphore.acquire)
-          }
-          .compile
-          .drain
-          .assertThrows[Err]
-      }
-
-      "when primary stream terminates, background stream is terminated" in forAll {
-        (s: Stream[Pure, Int]) =>
-          Stream
-            .eval(Semaphore[IO](0))
-            .flatMap { semaphore =>
-              val bg = Stream.repeatEval(IO(1) *> IO.sleep(50.millis)).onFinalize(semaphore.release)
-              val fg = s.delayBy[IO](25.millis)
-              fg.concurrently(bg)
-                .onFinalize(semaphore.acquire)
-            }
-            .compile
-            .drain
-            .assertNoException
-      }
-
-      "when background stream fails, primary stream fails even when hung" in forAll {
-        (s: Stream[Pure, Int]) =>
-          Stream
-            .eval(Deferred[IO, Unit])
-            .flatMap { gate =>
-              Stream(1)
-                .delayBy[IO](25.millis)
-                .append(s)
-                .concurrently(Stream.raiseError[IO](new Err))
-                .evalTap(_ => gate.get)
-            }
-            .compile
-            .drain
-            .assertThrows[Err]
-      }
-
-      "run finalizers of background stream and properly handle exception" in forAll {
-        s: Stream[Pure, Int] =>
-          Ref
-            .of[IO, Boolean](false)
-            .flatMap { runnerRun =>
-              Ref.of[IO, List[String]](Nil).flatMap { finRef =>
-                Deferred[IO, Unit].flatMap { halt =>
-                  def runner: Stream[IO, Unit] =
-                    Stream
-                      .bracket(runnerRun.set(true))(_ =>
-                        IO.sleep(
-                          100.millis
-                        ) >> // assure this inner finalizer always take longer run than `outer`
-                          finRef.update(_ :+ "Inner") >> // signal finalizer invoked
-                          IO.raiseError[Unit](new Err) // signal a failure
-                      ) >> // flag the concurrently had chance to start, as if the `s` will be empty `runner` may not be evaluated at all.
-                      Stream.eval_(halt.complete(())) // immediately interrupt the outer stream
-
-                  Stream
-                    .bracket(IO.unit)(_ => finRef.update(_ :+ "Outer"))
-                    .flatMap(_ => s.covary[IO].concurrently(runner))
-                    .interruptWhen(halt.get.attempt)
-                    .compile
-                    .drain
-                    .attempt
-                    .flatMap { r =>
-                      runnerRun.get.flatMap { runnerStarted =>
-                        finRef.get.flatMap { finalizers =>
-                          if (runnerStarted) IO {
-                            // finalizers shall be called in correct order and
-                            // exception shall be thrown
-                            assert(finalizers == List("Inner", "Outer"))
-                            assert(r.swap.toOption.get.isInstanceOf[Err])
-                          }
-                          else
-                            IO {
-                              // still the outer finalizer shall be run, but there is no failure in `s`
-                              assert(finalizers == List("Outer"))
-                              assert(r == Right(()))
-                            }
-                        }
-                      }
-                    }
-                }
-              }
-            }
-            .assertNoException
-      }
-    }
-
-    "debounce" in {
-      val delay = 200.milliseconds
-      (Stream(1, 2, 3) ++ Stream.sleep[IO](delay * 2) ++ Stream() ++ Stream(4, 5) ++ Stream
-        .sleep[IO](delay / 2) ++ Stream(6))
-        .debounce(delay)
-        .compile
-        .toList
-        .asserting(it => assert(it == List(3, 6)))
-    }
-
-    "delete" in forAll { (s: Stream[Pure, Int], idx0: PosZInt) =>
-      val v = s.toVector
-      val i = if (v.isEmpty) 0 else v(idx0 % v.size)
-      assert(s.delete(_ == i).toVector == v.diff(Vector(i)))
-    }
-
-    "drop" in forAll { (s: Stream[Pure, Int], negate: Boolean, n0: PosZInt) =>
-      val v = s.toVector
-      val n1 = if (v.isEmpty) 0 else n0 % v.size
-      val n = if (negate) -n1 else n1
-      assert(s.drop(n).toVector == s.toVector.drop(n))
-    }
-
-    "dropLast" in forAll { (s: Stream[Pure, Int]) =>
-      assert(s.dropLast.toVector == s.toVector.dropRight(1))
-    }
-
-    "dropLastIf" in forAll { (s: Stream[Pure, Int]) =>
-      assert(s.dropLastIf(_ => false).toVector == s.toVector)
-      assert(s.dropLastIf(_ => true).toVector == s.toVector.dropRight(1))
-    }
-
-    "dropRight" in forAll { (s: Stream[Pure, Int], negate: Boolean, n0: PosZInt) =>
-      val v = s.toVector
-      val n1 = if (v.isEmpty) 0 else n0 % v.size
-      val n = if (negate) -n1 else n1
-      assert(s.dropRight(n).toVector == v.dropRight(n))
-    }
-
-    "dropWhile" in forAll { (s: Stream[Pure, Int], n0: PosZInt) =>
-      val n = n0 % 20
-      val set = s.toVector.take(n).toSet
-      assert(s.dropWhile(set).toVector == s.toVector.dropWhile(set))
-    }
-
-    "dropThrough" in forAll { (s: Stream[Pure, Int], n0: PosZInt) =>
-      val n = n0 % 20
-      val set = s.toVector.take(n).toSet
-      val expected = {
-        val vec = s.toVector.dropWhile(set)
-        if (vec.isEmpty) vec else vec.tail
-      }
-      assert(s.dropThrough(set).toVector == expected)
-    }
-
-    "duration" in {
-      val delay = 200.millis
-      Stream
-        .emit(())
-        .append(Stream.eval(IO.sleep(delay)))
-        .zip(Stream.duration[IO])
-        .drop(1)
-        .map(_._2)
-        .compile
-        .toVector
-        .asserting { result =>
-          assert(result.size == 1)
-          val head = result.head
-          assert(head.toMillis >= (delay.toMillis - 5))
-        }
-    }
-
-    "either" in forAll { (s1: Stream[Pure, Int], s2: Stream[Pure, Int]) =>
-      val s1List = s1.toList
-      val s2List = s2.toList
-      s1.covary[IO].either(s2).compile.toList.asserting { result =>
-        assert(result.collect { case Left(i) => i } == s1List)
-        assert(result.collect { case Right(i) => i } == s2List)
-      }
-    }
-
-    "evalSeq" - {
-      "with List" in {
-        Stream
-          .evalSeq(IO(List(1, 2, 3)))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(1, 2, 3)))
-      }
-      "with Seq" in {
-        Stream.evalSeq(IO(Seq(4, 5, 6))).compile.toList.asserting(it => assert(it == List(4, 5, 6)))
-      }
-    }
-
-    "evalFilter" - {
-      "with effectful const(true)" in forAll { s: Stream[Pure, Int] =>
-        val s1 = s.toList
-        s.evalFilter(_ => IO.pure(true)).compile.toList.asserting(it => assert(it == s1))
-      }
-
-      "with effectful const(false)" in forAll { s: Stream[Pure, Int] =>
-        s.evalFilter(_ => IO.pure(false)).compile.toList.asserting(it => assert(it.isEmpty))
-      }
-
-      "with function that filters out odd elements" in {
-        Stream
-          .range(1, 10)
-          .evalFilter(e => IO(e % 2 == 0))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(2, 4, 6, 8)))
-      }
-    }
-
-    "evalFilterAsync" - {
-      "with effectful const(true)" in forAll { s: Stream[Pure, Int] =>
-        val s1 = s.toList
-        s.covary[IO]
-          .evalFilterAsync(5)(_ => IO.pure(true))
-          .compile
-          .toList
-          .asserting(it => assert(it == s1))
-      }
-
-      "with effectful const(false)" in forAll { s: Stream[Pure, Int] =>
-        s.covary[IO]
-          .evalFilterAsync(5)(_ => IO.pure(false))
-          .compile
-          .toList
-          .asserting(it => assert(it.isEmpty))
-      }
-
-      "with function that filters out odd elements" in {
-        Stream
-          .range(1, 10)
-          .evalFilterAsync[IO](5)(e => IO(e % 2 == 0))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(2, 4, 6, 8)))
-      }
-
-      "filters up to N items in parallel" in {
-        val s = Stream.range(0, 100)
-        val n = 5
-
-        (Semaphore[IO](n), SignallingRef[IO, Int](0)).tupled
-          .flatMap {
-            case (sem, sig) =>
-              val tested = s
-                .covary[IO]
-                .evalFilterAsync(n) { _ =>
-                  val ensureAcquired =
-                    sem.tryAcquire.ifM(
-                      IO.unit,
-                      IO.raiseError(new Throwable("Couldn't acquire permit"))
-                    )
-
-                  ensureAcquired.bracket(_ =>
-                    sig.update(_ + 1).bracket(_ => IO.sleep(10.millis))(_ => sig.update(_ - 1))
-                  )(_ => sem.release) *>
-                    IO.pure(true)
-                }
-
-              sig.discrete
-                .interruptWhen(tested.drain)
-                .fold1(_.max(_))
-                .compile
-                .lastOrError
-                .product(sig.get)
-          }
-          .asserting(it => assert(it == ((n, 0))))
-      }
-    }
-
-    "evalFilterNot" - {
-      "with effectful const(true)" in forAll { s: Stream[Pure, Int] =>
-        val s1 = s.toList
-        s.evalFilterNot(_ => IO.pure(false)).compile.toList.asserting(it => assert(it == s1))
-      }
-
-      "with effectful const(false)" in forAll { s: Stream[Pure, Int] =>
-        s.evalFilterNot(_ => IO.pure(true)).compile.toList.asserting(it => assert(it.isEmpty))
-      }
-
-      "with function that filters out odd elements" in {
-        Stream
-          .range(1, 10)
-          .evalFilterNot(e => IO(e % 2 == 0))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(1, 3, 5, 7, 9)))
-      }
-    }
-
-    "evalFilterNotAsync" - {
-      "with effectful const(true)" in forAll { s: Stream[Pure, Int] =>
-        s.covary[IO]
-          .evalFilterNotAsync(5)(_ => IO.pure(true))
-          .compile
-          .toList
-          .asserting(it => assert(it.isEmpty))
-      }
-
-      "with effectful const(false)" in forAll { s: Stream[Pure, Int] =>
-        val s1 = s.toList
-        s.covary[IO]
-          .evalFilterNotAsync(5)(_ => IO.pure(false))
-          .compile
-          .toList
-          .asserting(it => assert(it == s1))
-      }
-
-      "with function that filters out odd elements" in {
-        Stream
-          .range(1, 10)
-          .evalFilterNotAsync[IO](5)(e => IO(e % 2 == 0))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(1, 3, 5, 7, 9)))
-      }
-
-      "filters up to N items in parallel" in {
-        val s = Stream.range(0, 100)
-        val n = 5
-
-        (Semaphore[IO](n), SignallingRef[IO, Int](0)).tupled
-          .flatMap {
-            case (sem, sig) =>
-              val tested = s
-                .covary[IO]
-                .evalFilterNotAsync(n) { _ =>
-                  val ensureAcquired =
-                    sem.tryAcquire.ifM(
-                      IO.unit,
-                      IO.raiseError(new Throwable("Couldn't acquire permit"))
-                    )
-
-                  ensureAcquired.bracket(_ =>
-                    sig.update(_ + 1).bracket(_ => IO.sleep(10.millis))(_ => sig.update(_ - 1))
-                  )(_ => sem.release) *>
-                    IO.pure(false)
-                }
-
-              sig.discrete
-                .interruptWhen(tested.drain)
-                .fold1(_.max(_))
-                .compile
-                .lastOrError
-                .product(sig.get)
-          }
-          .asserting(it => assert(it == ((n, 0))))
-      }
-    }
-
-    "evalMapAccumulate" in forAll { (s: Stream[Pure, Int], m: Int, n0: PosInt) =>
-      val sVector = s.toVector
-      val n = n0 % 20 + 1
-      val f = (_: Int) % n == 0
-      val r = s.covary[IO].evalMapAccumulate(m)((s, i) => IO.pure((s + i, f(i))))
-      r.map(_._1).compile.toVector.asserting(it => assert(it == sVector.scanLeft(m)(_ + _).tail))
-      r.map(_._2).compile.toVector.asserting(it => assert(it == sVector.map(f)))
-    }
-
-    "evalMapFilter" - {
-      "with effectful optional identity function" in forAll { s: Stream[Pure, Int] =>
-        val s1 = s.toList
-        s.evalMapFilter(n => IO.pure(n.some)).compile.toList.asserting(it => assert(it == s1))
-      }
-
-      "with effectful constant function that returns None for any element" in forAll {
-        s: Stream[Pure, Int] =>
-          s.evalMapFilter(_ => IO.pure(none[Int]))
-            .compile
-            .toList
-            .asserting(it => assert(it.isEmpty))
-      }
-
-      "with effectful function that filters out odd elements" in {
-        Stream
-          .range(1, 10)
-          .evalMapFilter(e => IO.pure(e.some.filter(_ % 2 == 0)))
-          .compile
-          .toList
-          .asserting(it => assert(it == List(2, 4, 6, 8)))
-      }
-    }
-
-    "evalScan" in forAll { (s: Stream[Pure, Int], n: String) =>
-      val sVector = s.toVector
-      val f: (String, Int) => IO[String] = (a: String, b: Int) => IO.pure(a + b)
-      val g = (a: String, b: Int) => a + b
-      s.covary[IO]
-        .evalScan(n)(f)
-        .compile
-        .toVector
-        .asserting(it => assert(it == sVector.scanLeft(n)(g)))
-    }
-
-    "every" in {
-      flickersOnTravis
-      type BD = (Boolean, FiniteDuration)
-      def durationSinceLastTrue[F[_]]: Pipe[F, BD, BD] = {
-        def go(lastTrue: FiniteDuration, s: Stream[F, BD]): Pull[F, BD, Unit] =
-          s.pull.uncons1.flatMap {
-            case None => Pull.done
-            case Some((pair, tl)) =>
-              pair match {
-                case (true, d) =>
-                  Pull.output1((true, d - lastTrue)) >> go(d, tl)
-                case (false, d) =>
-                  Pull.output1((false, d - lastTrue)) >> go(lastTrue, tl)
-              }
-          }
-        s => go(0.seconds, s).stream
-      }
-
-      val delay = 20.millis
-      val draws = (600.millis / delay).min(50) // don't take forever
-
-      val durationsSinceSpike = Stream
-        .every[IO](delay)
-        .map(d => (d, System.nanoTime.nanos))
-        .take(draws.toInt)
-        .through(durationSinceLastTrue)
-
-      (IO.shift >> durationsSinceSpike.compile.toVector).unsafeToFuture().map { result =>
-        val list = result.toList
-        withClue("every always emits true first")(assert(list.head._1))
-        withClue(s"true means the delay has passed: ${list.tail}") {
-          assert(list.tail.filter(_._1).map(_._2).forall(_ >= delay))
-        }
-        withClue(s"false means the delay has not passed: ${list.tail}") {
-          assert(list.tail.filterNot(_._1).map(_._2).forall(_ <= delay))
-        }
-      }
-    }
-
     "exists" in forAll { (s: Stream[Pure, Int], n0: PosInt) =>
       val n = n0 % 20 + 1
       val f = (i: Int) => i % n == 0
